@@ -2,10 +2,9 @@ import React, { useState } from "react";
 import useSWR from "swr";
 
 // material ui
-import { makeStyles } from "@material-ui/core/styles";
-import Grid from "@material-ui/core/Grid";
-import ListItem from "@material-ui/core/ListItem";
-import ListItemIcon from "@material-ui/core/ListItemIcon";
+import Grid from "@mui/material/Grid";
+import ListItem from "@mui/material/ListItem";
+import ListItemIcon from "@mui/material/ListItemIcon";
 import IconButton from "@mui/material/IconButton";
 import ExpandLess from "@mui/icons-material/ExpandLess";
 import ExpandMore from "@mui/icons-material/ExpandMore";
@@ -22,28 +21,27 @@ import Expander from "./Expander";
 
 //utils
 import useSWRAbstract from "../util/swr"
-
-const useStyles = makeStyles((theme) => ({
-  margin: {
-    margin: theme.spacing(1),
-    padding: 5,
-  },
-  extendedIcon: {
-    marginRight: theme.spacing(1),
-  },
-  floater: {
-    position: "absolute",
-    top: "10px",
-  },
-  itemAlign: {
-    display: "flex",
-    alignItems: "center",
-    flexWrap: "wrap",
-  },
-}));
+// 
+// const useStyles = makeStyles((theme) => ({
+//   margin: {
+//     margin: theme.spacing(1),
+//     padding: 5,
+//   },
+//   extendedIcon: {
+//     marginRight: theme.spacing(1),
+//   },
+//   floater: {
+//     position: "absolute",
+//     top: "10px",
+//   },
+//   itemAlign: {
+//     display: "flex",
+//     alignItems: "center",
+//     flexWrap: "wrap",
+//   },
+// }));
 
 export default function PostListItem(props) {
-  const classes = useStyles();
   const { post, post_loading, post_error } = useSWRAbstract("post", `/api/posts/${props.id}`);
   const dispatch = useDispatch();
   const [open, setOpen] = useState(false);
