@@ -21,7 +21,7 @@ def session(db):
 	curr_session_id = str(res.inserted_id)
 	yield curr_session_id # this is where the testing happens
 	# remove the session
-	db.sessions.remove({'_id': res.inserted_id})
+	db.sessions.delete_one({'_id': res.inserted_id})
 
 @pytest.fixture
 def user(db):
@@ -31,17 +31,16 @@ def user(db):
 	curr_user_id = str(res.inserted_id)
 	yield 'test' # this is where the testing happens
 	# Remove the user
-	db.users.remove({'_id': res.inserted_id})
+	db.users.delete_one({'_id': res.inserted_id})
 
 
 # ! Test cases for initialize_teleoscope
-
 # Case 1: Invalid session id - should throw an exception
 def test_initialize_teleoscope_dummy_label(db):
 	with pytest.raises(Exception):
 		tasks.initialize_teleoscope((), label = "test label", session_id = 1010)
 	# delete test label teleoscope from table
-	db.teleoscopes.remove({ 'label': "test label" })
+	db.teleoscopes.delete_one({ 'label': "test label" })
 
 # Case 2: Valid label, and session id
 def test_initialize_teleoscope_valid_label(db, session):
@@ -50,7 +49,7 @@ def test_initialize_teleoscope_valid_label(db, session):
 	except Exception as e:
 		pytest.fail(e)
 	finally:
-		db.teleoscopes.remove({ 'label': "test label" })
+		db.teleoscopes.delete_one({ 'label': "test label" })
 
 
 # ! Test cases for initialize_session
@@ -62,6 +61,53 @@ def test_initialize_session_valid_username(db, user):
 			pytest.fail(e)
 		finally:
 			# Remove session
-			db.sessions.remove({'username': 'test'})
+			db.sessions.delete_one({'username': 'test'})
 	
+# ! Test cases for save_teleoscope_state
+# Case 1: Invalid Teleoscope id
+def test_save_teleoscope_state_invalid_id():
+	with pytest.raises(Exception):
+		tasks.save_teleoscope_state((), history_object = {"_id": '42', "history_item": {}})
+
+# ! Test cases for save_UI_state
+# Case 1: Invalid session id
+def test_save_UI_state_invalid_session_id():
+	with pytest.raises(Exception):
+		tasks.save_UI_state((), history_object = {"_id": '42', "history_item": {}})
+
+# ! Test cases for add_group
+# Case 1: Invalid session id
+def test_add_group_invalid_session_id():
+	with pytest.raises(Exception):
+		tasks.add_group((), label = "test label", color = "test color", session_id = '42')
+
+# ! Test cases for add_post_to_group
+# Case 1: Invalid group id
+def test_add_post_to_group_invalid_group_id():
+	with pytest.raises(Exception):
+		tasks.add_post_to_group((), post_id = '42', group_id = '42')
+
+# ! Test cases for remove_post_from_group
+# Case 2: Invalid group id
+def test_remove_post_from_group_invalid_group_id():
+	with pytest.raises(Exception):
+		tasks.remove_post_from_group((), post_id = '42', group_id = '42')
+
+# ! Test cases for update group_label
+# Case 2: Invalid group id
+def test_update_group_label_invalid_group_id():
+	with pytest.raises(Exception):
+		tasks.update_group_label((), group_id = '42', label = 'test label')
+
+# ! Test cases for add_note
+# Case 1: Invalid post_id
+def test_add_note_invalid_post_id():
+	with pytest.raises(Exception):
+		tasks.add_note((), post_id = '42')
+
+# ! Test cases for update_note
+# Case 1: Invalid post_id
+def test_update_note_invalid_post_id():
+	with pytest.raises(Exception):
+		tasks.update_note((), post_id = '42', content={})
 
