@@ -1,7 +1,6 @@
 module.exports = {
-  preset: 'ts-jest',
+  // preset: 'ts-jest',
   transform: {
-    '^.+\\.(ts|tsx)?$': 'ts-jest',
     "^.+\\.(js|jsx)$": "babel-jest",
   }
 };
